:recycle: convert to functional
